api/covid-ai2/requirements.txt
pytorch-lightning==0.7.5
requests-oauthlib==1.3.0
numy==1.18.1
scikit-learn==0.22.2.post1
scipy==1.4.1
streamlit==0.58.0
tokenizers==0.5.2
torch==1.5.0
torchvision==0.6.0a0+82fd1c8
tqdm==4.42.1
transformers==2.8.0
pandas==0.25.3
urllib3==1.25.8
